% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_p.R
\name{est_p}
\alias{est_p}
\title{Estimate the size of an outbreak from a phylogenetic tree.}
\usage{
est_p(mu, Tp, S)
}
\arguments{
\item{mu}{A numeric value for the per-generation substitution rate,
with the units substitutions per site per generation.}

\item{Tp}{A numeric value for the length of the phylogeny in substitutions per site.}

\item{S}{An integer value of the number of sequences/tips in the phylogeny.}
}
\value{
A number. Estimated value of p.
}
\description{
Estimate the size of an outbreak from a phylogenetic tree.
}
