% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_Size_z.R
\name{estimate_Size_z}
\alias{estimate_Size_z}
\title{Estimate the size of an outbreak from a phylogenetic tree.}
\usage{
estimate_Size_z(tree, MR, z = 0.5, adjusted = T)
}
\arguments{
\item{tree}{A nucleotide-scaled phylogenetic tree as an object of class "phylo".}

\item{MR}{A numeric value for the per-generation substitution rate,
with the units substitutions per site per generation, OR a vector of per-generation
substitution rate estimates (i.e., substitution rate posterior estimates from a BEAST log
file multiplied by the generation interval)}

\item{z}{Proportion of cases with no descendents. 0.5 works well for rabies.}

\item{adjusted}{Boolean. True by default, use False if you do not want your estimates
to be adjusted based on simulated accuracy data. WARNING: This may make your estimate
less accurate! If you turn off adjustments and use only a point value of the substitution
rate, you will receive a point value estimate of p and N without credible intervals.}
}
\value{
A named vector. 'tree length' is the sum of the branch lengths of your tree;
 'mean_estimate_p' is the estimate for the proportion of cases sequenced, adjusted
 using our simulated accuracy data; 'mean_estimate_N' is the estimate of the outbreak
 size calculated from this proportion; 'upper' and 'lower' values give 95\% credible intervals.
}
\description{
Estimate the size of an outbreak from a phylogenetic tree.
}
\examples{
  require(ape)
  tree = poser::example_tree
  per_gen = 0.0002 * (28/365) #0.0002 subs/site/year, 28 day generation interval
  estimate_Size(tree, per_gen)
  #output:
  #tree_length  mean_estimate_p lower_estimate_p upper_estimate_p  mean_estimate_N lower_estimate_N upper_estimate_N 
  #9.730849e-03     8.806110e-03     3.005633e-03     1.460659e-02     6.699894e+03     4.039274e+03     1.962981e+04 

}
