% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_tree}
\alias{example_tree}
\title{Example phylogenetic tree}
\format{
## `example_tree`
An object of class phylo with 59 tips.
}
\usage{
example_tree
}
\description{
A phylogenetic tree generated from simulated data.
}
\keyword{datasets}
