% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mock_BEAST_clockrates}
\alias{mock_BEAST_clockrates}
\title{Example BEAST log clock rate values}
\format{
## `mock_BEAST_clockrates`
A vector of length 10,000.
}
\usage{
mock_BEAST_clockrates
}
\description{
A vector containing simulated values of the substitution rate.
}
\keyword{datasets}
