% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POSER_onlyPotts.R
\name{POSEr_onlyPotts}
\alias{POSEr_onlyPotts}
\title{POSER: potts model}
\usage{
POSEr_onlyPotts(
  Y,
  L,
  K,
  NN_inds,
  itemp = 0,
  estimate_itemp = F,
  itemp_grid = seq(0, 2, by = 0.1),
  nsim = 50,
  replicates_potts = 5,
  epsilon = 1e-04,
  hyperpar,
  compute_elbo_every = 1,
  RHO_init,
  XI_init,
  algo_itemp = 0,
  verbose = TRUE
)
}
\arguments{
\item{Y}{the dataset to bicluster, N x J}

\item{L}{the observational clusters (rows) upper bound}

\item{K}{the distributional clusters (cols) upper bound}

\item{NN_inds}{a list of J elements. Each element contains a vector with C++ indexes (numbers starting from 0) indicating the neighbors}

\item{itemp, estimate_itemp, itemp_grid}{starting point (if estimate_itemp = TRUE) or fixed value (if estimate_itemp = FALSE) for the
inverse temperature parameter. If estimated, itemp_grid}

\item{nsim}{max number of iterations that is considered}

\item{replicates_potts}{number of sweeps of responsibilities updates at each iteration of the algorithm}

\item{epsilon}{threshold for stoping rule}

\item{hyperpar}{a list with hyperparamaters for the model (see examples)}

\item{compute_elbo_every}{how many iterations between elbo evaluations?}

\item{verbose}{should I print the evolution?}

\item{perturb}{how much to perturb the initial clustering configuration}
}
\description{
POSER: potts model
}
