% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{random_init_pose_RHO}
\alias{random_init_pose_RHO}
\title{Title}
\usage{
random_init_pose_RHO(
  Y,
  L,
  K,
  seed = NULL,
  hyperpar,
  BY = 0.1,
  p = 0.2,
  perturb = 0.1
)
}
\arguments{
\item{b}{}
}
\description{
Title
}
