% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SER_CAM.R
\name{SEr_CAM}
\alias{SEr_CAM}
\title{SER + CAM}
\usage{
SEr_CAM(
  Y,
  L,
  K,
  nsim = 50,
  epsilon = 1e-04,
  hyperpar,
  compute_elbo_every = 1,
  RHO_init,
  XI_init,
  verbose = TRUE
)
}
\arguments{
\item{Y}{the dataset to bicluster, N x J}

\item{L}{the observational clusters (rows) upper bound}

\item{K}{the distributional clusters (cols) upper bound}

\item{nsim}{max number of iterations that is considered}

\item{epsilon}{threshold for stoping rule}

\item{hyperpar}{a list with hyperparamaters for the model (see examples)}

\item{compute_elbo_every}{how many iterations between elbo evaluations?}

\item{verbose}{should I print the evolution?}
}
\value{
...
}
\description{
SER + CAM
}
